Change MAG_SetInterruptThreshold argument to float
- Change accepted argument from uint16_t to float
- Return error when out of bounds

Using float makes more sense because the magnetometer output values are also float.

+++ src/SparkFunLSM303C.h
@@ -29,7 +29,7 @@ public:
 
     //Magnetometer Interrupts
     SparkFunIMU_status_t MAG_EnableInterrupt(bool, LSM303C_AXIS_t);
-    SparkFunIMU_status_t MAG_SetInterruptThreshold(uint16_t);
+    SparkFunIMU_status_t MAG_SetInterruptThreshold(float);
     SparkFunIMU_status_t MAG_ResetInterruptConfig();
     SparkFunIMU_status_t MAG_ResetInterruptLatch();
     SparkFunIMU_status_t MAG_ReadIntReg();
